+++ Sources/NablaMessagingUI/Scenes/Conversation/UI/ConversationViewController.swift
@@ -95,16 +95,26 @@ final class ConversationViewController: UIViewController, ConversationViewContra
 
     func displayImageDetail(for image: ImageFile) {
         let viewController = ImageDetailViewController()
+
         let presenter = ImageDetailPresenterImpl(viewContract: viewController, image: image)
         viewController.presenter = presenter
-        show(viewController, sender: nil)
+
+        let navigationController = UINavigationController(rootViewController: viewController).nabla.withOpaqueNavigationBarBackground()
+        navigationController.modalPresentationStyle = .fullScreen
+
+        present(navigationController, animated: true)
     }
 
     func displayDocumentDetail(for document: DocumentFile) {
         let viewController = DocumentDetailViewController()
+        
         let presenter = DocumentDetailPresenterImpl(viewContract: viewController, document: document)
         viewController.presenter = presenter
-        show(viewController, sender: nil)
+
+        let navigationController = UINavigationController(rootViewController: viewController).nabla.withOpaqueNavigationBarBackground()
+        navigationController.modalPresentationStyle = .fullScreen
+        
+        present(navigationController, animated: true)
     }
     
     func displayDocumentScanner() {